+++ lib/ravelin/authentication_mechanism.rb
@@ -16,5 +16,9 @@ module Ravelin
     def social=(mechanism)
       @social = Ravelin::AuthenticationMechanisms::Social.new(mechanism)
     end
+
+    def sms_code=(mechanism)
+      @sms_code = Ravelin::AuthenticationMechanisms::SmsCode.new(mechanism)
+    end
   end
 end